Fix training correctness across detection, gradients, and validation
Correct YOLO anchor matching by avoiding integer division, and replace bounding-envelope overlap calculations with true set-union IoU. Keep NMS isolated per batch image and add overlap, cross-batch, and metric regressions.

Preserve gradients through branched networks by accumulating CUDA convolution, pooling, connected, batch-normalization, dropout, softmax, shortcut, and CPU convolution paths. Correct shortcut scaling, dropout propagation, upsample lifecycle handling, and reject invalid route and max-pool configurations.

Replace cuDNN max pooling with a bounds-safe CUDA implementation that supports Darknet's asymmetric edge padding for the final YOLO-tiny stride-1 pool. Track argmax indexes safely and atomically accumulate backward gradients, with a CUDA regression covering forward geometry and backward accumulation.

Normalize optimizer updates by the effective batch across subdivisions, apply Adam weight decay only to weights, use optimizer-update steps for bias correction, and check all cuBLAS operations. Save and restore Adam moments and step counters in optimizer sidecars so resumed training retains optimizer state.

Make validation deterministic and augmentation-free, bind validation ground truth to the active model batch, cap validation passes to available data, calculate confidence-ranked per-class AP, and include ghost predictions and missed objects in micro-F1.

Update the tiny VOC training configuration to train-200/val-10 and register the new CUDA max-pool source and tests.

+++ include/YoloLayer.h
@@ -415,8 +415,8 @@ void YoloLayer<D>::processObjects(int b)
 
         for (auto n = 0; n < numAnchors_; ++n) {
             DarkBox pred;
-            pred.w() = anchors_[2 * n] / this->model().width();
-            pred.h() = anchors_[2 * n + 1] / this->model().height();
+            pred.w() = static_cast<float>(anchors_[2 * n]) / this->model().width();
+            pred.h() = static_cast<float>(anchors_[2 * n + 1]) / this->model().height();
 
             auto iou = pred.iou(truthShift);
             if (iou > bestIoU) {
@@ -682,4 +682,3 @@ using CudaYolo = YoloLayer<Device::CUDA>;
 #include "cuda/YoloLayer.h"
 
 #endif // USE_CUDA
-
